Refactor and remove utils.go
