+++ functionaltests/steps.go
@@ -20,10 +20,14 @@ package functionaltests
 import (
 	"context"
 	"fmt"
+	"maps"
+	"slices"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/require"
+
+	"github.com/elastic/go-elasticsearch/v8/typedapi/types"
 
 	"github.com/elastic/apm-server/functionaltests/internal/asserts"
 	"github.com/elastic/apm-server/functionaltests/internal/ecclient"
@@ -168,14 +172,22 @@ func (c createStep) Step(t *testing.T, ctx context.Context, e *testStepEnv, _ te
 //
 // NOTE: Only works for versions >= 8.0.
 type ingestStep struct {
+	// CheckDataStream is used to check all data streams have the same expected values.
+	// For checking each data stream individually, use CheckIndividualDataStream.
 	CheckDataStream asserts.CheckDataStreamsWant
-	// IgnoreDataStreams are the data streams to be ignored in assertions.
-	// The data stream names can contain '%s' to indicate namespace.
-	IgnoreDataStreams []string
+
 	// CheckIndividualDataStream is used to check the data streams individually
 	// instead of as a whole using CheckDataStream.
 	// The data stream names can contain '%s' to indicate namespace.
 	CheckIndividualDataStream map[string]asserts.CheckDataStreamIndividualWant
+
+	// IgnoreDataStreams are the data streams to be ignored in assertions.
+	// The data stream names can contain '%s' to indicate namespace.
+	IgnoreDataStreams []string
+
+	// SkipIndices skips checking the indices within data streams,
+	// i.e. only data streams are checked.
+	SkipIndices bool
 }
 
 func (i ingestStep) Step(t *testing.T, ctx context.Context, e *testStepEnv, previousRes testStepResult) testStepResult {
@@ -198,9 +210,11 @@ func (i ingestStep) Step(t *testing.T, ctx context.Context, e *testStepEnv, prev
 	dataStreams := getAPMDataStreams(t, ctx, e.esc, ignoreDS...)
 	if i.CheckIndividualDataStream != nil {
 		expected := formatAllMap(i.CheckIndividualDataStream, e.dsNamespace)
-		asserts.CheckDataStreamsIndividually(t, expected, dataStreams)
+		asserts.CheckDataStreamsIndividually(t, expected, dataStreams,
+			asserts.SkipIndices(i.SkipIndices))
 	} else {
-		asserts.CheckDataStreams(t, i.CheckDataStream, dataStreams)
+		asserts.CheckDataStreams(t, i.CheckDataStream, dataStreams,
+			asserts.SkipIndices(i.SkipIndices))
 	}
 
 	return testStepResult{DSDocCount: dsDocCount}
@@ -230,15 +244,25 @@ func formatAllMap[T any](m map[string]T, s string) map[string]T {
 //
 // NOTE: Only works from versions >= 8.0.
 type upgradeStep struct {
-	NewVersion      ecclient.StackVersion
+	// NewVersion is the version to upgrade into.
+	NewVersion ecclient.StackVersion
+
+	// CheckDataStream is used to check all data streams have the same expected values.
+	// For checking each data stream individually, use CheckIndividualDataStream.
 	CheckDataStream asserts.CheckDataStreamsWant
-	// IgnoreDataStreams are the data streams to be ignored in assertions.
-	// The data stream names can contain '%s' to indicate namespace.
-	IgnoreDataStreams []string
+
 	// CheckIndividualDataStream is used to check the data streams individually
 	// instead of as a whole using CheckDataStream.
 	// The data stream names can contain '%s' to indicate namespace.
 	CheckIndividualDataStream map[string]asserts.CheckDataStreamIndividualWant
+
+	// IgnoreDataStreams are the data streams to be ignored in assertions.
+	// The data stream names can contain '%s' to indicate namespace.
+	IgnoreDataStreams []string
+
+	// SkipIndices skips checking the indices within data streams,
+	// i.e. only data streams are checked.
+	SkipIndices bool
 }
 
 func (u upgradeStep) Step(t *testing.T, ctx context.Context, e *testStepEnv, previousRes testStepResult) testStepResult {
@@ -265,9 +289,11 @@ func (u upgradeStep) Step(t *testing.T, ctx context.Context, e *testStepEnv, pre
 	dataStreams := getAPMDataStreams(t, ctx, e.esc, ignoreDS...)
 	if u.CheckIndividualDataStream != nil {
 		expected := formatAllMap(u.CheckIndividualDataStream, e.dsNamespace)
-		asserts.CheckDataStreamsIndividually(t, expected, dataStreams)
+		asserts.CheckDataStreamsIndividually(t, expected, dataStreams,
+			asserts.SkipIndices(u.SkipIndices))
 	} else {
-		asserts.CheckDataStreams(t, u.CheckDataStream, dataStreams)
+		asserts.CheckDataStreams(t, u.CheckDataStream, dataStreams,
+			asserts.SkipIndices(u.SkipIndices))
 	}
 
 	return testStepResult{DSDocCount: dsDocCount}
@@ -279,7 +305,12 @@ func (u upgradeStep) Step(t *testing.T, ctx context.Context, e *testStepEnv, pre
 //
 // The output of this step is the previous step's result.
 type checkErrorLogsStep struct {
-	ESErrorLogsIgnored  esErrorLogs
+	// ESErrorLogsIgnored are the error logs query from Elasticsearch that are
+	// to be ignored.
+	ESErrorLogsIgnored esErrorLogs
+
+	// APMErrorLogsIgnored are the error logs query from APM Server that are
+	// to be ignored.
 	APMErrorLogsIgnored apmErrorLogs
 }
 
@@ -308,4 +339,95 @@ type customStep struct {
 
 func (c customStep) Step(t *testing.T, ctx context.Context, e *testStepEnv, previousRes testStepResult) testStepResult {
 	return c.Func(t, ctx, e, previousRes)
+}
+
+// expectedDataStreamsIngest represent the expected number of ingested document
+// after a single run of ingest.
+//
+// NOTE: The aggregation data streams have negative counts, because they are
+// expected to appear but the document counts should not be asserted.
+func expectedDataStreamsIngest(namespace string) esclient.DataStreamsDocCount {
+	return map[string]int{
+		fmt.Sprintf("traces-apm-%s", namespace):                     15013,
+		fmt.Sprintf("metrics-apm.app.opbeans_python-%s", namespace): 1437,
+		fmt.Sprintf("metrics-apm.internal-%s", namespace):           1351,
+		fmt.Sprintf("logs-apm.error-%s", namespace):                 364,
+		// Ignore aggregation data streams.
+		fmt.Sprintf("metrics-apm.service_destination.1m-%s", namespace): -1,
+		fmt.Sprintf("metrics-apm.service_transaction.1m-%s", namespace): -1,
+		fmt.Sprintf("metrics-apm.service_summary.1m-%s", namespace):     -1,
+		fmt.Sprintf("metrics-apm.transaction.1m-%s", namespace):         -1,
+	}
+}
+
+// emptyDataStreamsIngest represent an empty ingestion.
+// It is useful for asserting that the document count did not change after an operation.
+//
+// NOTE: The aggregation data streams have negative counts, because they
+// are expected to appear but the document counts should not be asserted.
+func emptyDataStreamsIngest(namespace string) esclient.DataStreamsDocCount {
+	return map[string]int{
+		fmt.Sprintf("traces-apm-%s", namespace):                     0,
+		fmt.Sprintf("metrics-apm.app.opbeans_python-%s", namespace): 0,
+		fmt.Sprintf("metrics-apm.internal-%s", namespace):           0,
+		fmt.Sprintf("logs-apm.error-%s", namespace):                 0,
+		// Ignore aggregation data streams.
+		fmt.Sprintf("metrics-apm.service_destination.1m-%s", namespace): -1,
+		fmt.Sprintf("metrics-apm.service_transaction.1m-%s", namespace): -1,
+		fmt.Sprintf("metrics-apm.service_summary.1m-%s", namespace):     -1,
+		fmt.Sprintf("metrics-apm.transaction.1m-%s", namespace):         -1,
+	}
+}
+
+func allDataStreams(namespace string) []string {
+	return slices.Collect(maps.Keys(expectedDataStreamsIngest(namespace)))
+}
+
+func sliceToSet[T comparable](s []T) map[T]bool {
+	m := make(map[T]bool)
+	for _, ele := range s {
+		m[ele] = true
+	}
+	return m
+}
+
+// getAPMDataStreams get all APM related data streams.
+func getAPMDataStreams(t *testing.T, ctx context.Context, esc *esclient.Client, ignoreDS ...string) []types.DataStream {
+	t.Helper()
+	dataStreams, err := esc.GetDataStream(ctx, "*apm*")
+	require.NoError(t, err)
+
+	ignore := sliceToSet(ignoreDS)
+	return slices.DeleteFunc(dataStreams, func(ds types.DataStream) bool {
+		return ignore[ds.Name]
+	})
+}
+
+// getDocCountPerDS retrieves document count per data stream for versions >= 8.0.
+func getDocCountPerDS(t *testing.T, ctx context.Context, esc *esclient.Client, ignoreDS ...string) esclient.DataStreamsDocCount {
+	t.Helper()
+	count, err := esc.APMDSDocCount(ctx)
+	require.NoError(t, err)
+
+	ignore := sliceToSet(ignoreDS)
+	maps.DeleteFunc(count, func(ds string, _ int) bool {
+		return ignore[ds]
+	})
+	return count
+}
+
+// getDocCountPerDS retrieves document count per data stream for versions < 8.0.
+func getDocCountPerDSV7(t *testing.T, ctx context.Context, esc *esclient.Client, namespace string) esclient.DataStreamsDocCount {
+	t.Helper()
+	count, err := esc.APMDSDocCountV7(ctx, namespace)
+	require.NoError(t, err)
+	return count
+}
+
+// getDocCountPerIndexV7 retrieves document count per index for versions < 8.0.
+func getDocCountPerIndexV7(t *testing.T, ctx context.Context, esc *esclient.Client) esclient.IndicesDocCount {
+	t.Helper()
+	count, err := esc.APMIdxDocCountV7(ctx)
+	require.NoError(t, err)
+	return count
 }
